requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic[email]
passlib[argon2]
python-jose
sib-api-v3-sdk
pynacl
